Refactor dataUtils
