+++ src/utilities/data-utils/index.ts
@@ -1,0 +1,7 @@
+import {initializeFormData} from './initialize-form-data';
+import {initializeFormValidation} from './initialize-form-validation';
+
+export const dataUtils = {
+    initializeFormData,
+    initializeFormValidation,
+};
